requirements.txt
python-dotenv
sqlalchemy
werkzeug
flask
flask-sqlalchemy
flask-login